+++ packages/injective-ui/lib/tailwind-preset.cjs
@@ -5,49 +5,6 @@ module.exports = {
       'glow-lg': '0px 0px 20px rgba(255, 255, 255, 0.2)'
     },
 
-    colors: {
-      black: '#14151A',
-      transparent: 'transparent',
-      white: '#ffffff',
-
-      primary: {
-        500: '#00F2FE'
-      },
-
-      accent: {
-        500: '#EA6D51'
-      },
-
-      gray: {
-        300: '#F2F2F2',
-        600: '#A1A1A3',
-        700: '#727376',
-        750: '#434448',
-        800: '#232C39',
-        900: '#171D29'
-      },
-
-      yellow: {
-        600: '#F5CD6E'
-      },
-
-      green: {
-        400: '#3EE8AF',
-        600: '#0EE29B'
-      },
-
-      red: {
-        400: '#FD6384',
-        600: '#F3164D'
-      },
-
-      blue: {
-        300: '#0076A5',
-        400: '#80C2FF',
-        600: '#0082FA'
-      }
-    },
-
     screens: {
       xs: '475px',
       sm: '640px',
@@ -79,7 +36,9 @@ module.exports = {
         20: '5rem',
         40: '10rem',
         xs: '20rem',
-        sm: '24rem'
+        sm: '24rem',
+        md: '28rem',
+        lg: '32rem'
       },
       maxWidth: {
         '90%': '90%'
@@ -93,6 +52,53 @@ module.exports = {
         base: ['1rem', { lineHeight: '1rem' }],
         xl: ['1.25rem', { lineHeight: '1.25rem' }],
         '3xl': ['2rem', { lineHeight: '2rem' }]
+      },
+      colors: {
+        black: '#14151A',
+        transparent: 'transparent',
+        white: '#ffffff',
+
+        primary: {
+          500: '#00F2FE'
+        },
+
+        accent: {
+          500: '#EA6D51'
+        },
+
+        gray: {
+          300: '#F2F2F2',
+          600: '#A1A1A3',
+          700: '#727376',
+          750: '#434448',
+          800: '#232C39',
+          900: '#171D29'
+        },
+
+        yellow: {
+          600: '#F5CD6E'
+        },
+
+        green: {
+          400: '#3EE8AF',
+          600: '#0EE29B'
+        },
+
+        red: {
+          400: '#FD6384',
+          500: '#F3164D',
+          600: '#F3164D'
+        },
+
+        orange: {
+          400: '#F7931A'
+        },
+
+        blue: {
+          300: '#0076A5',
+          400: '#80C2FF',
+          600: '#0082FA'
+        }
       }
     }
   }
