+++ packages/injective-ui/package.json
@@ -35,9 +35,10 @@
     "floating-vue": "^2.0.0-beta.20"
   },
   "devDependencies": {
-    "@injectivelabs/eslint-config": "^1.0.2-alpha.0",
-    "@injectivelabs/prettier-config": "^1.0.0-alpha.0",
-    "@injectivelabs/stylelint-config": "^1.0.1-alpha.0",
+    "@injectivelabs/notifications": "*",
+    "@injectivelabs/eslint-config": "*",
+    "@injectivelabs/prettier-config": "*",
+    "@injectivelabs/stylelint-config": "*",
     "@nuxt/module-builder": "latest",
     "@nuxt/schema": "^3.0.0-rc.12",
     "@nuxtjs/eslint-config-typescript": "latest",
