+++ packages/injective-ui/tsconfig.json
@@ -2,7 +2,7 @@
   "extends": "./playground/.nuxt/tsconfig.json",
   "compilerOptions": {
     "types": [
-      "@vueuse/nuxt",
+      "@vueuse/nuxt"
     ]
   }
 }
