+++ packages/notifications/package.json
@@ -31,9 +31,9 @@
     "@nuxt/kit": "^3.0.0-rc.12"
   },
   "devDependencies": {
-    "@injectivelabs/eslint-config": "*",
-    "@injectivelabs/prettier-config": "*",
-    "@injectivelabs/stylelint-config": "*",
+    "@injectivelabs/eslint-config": "latest",
+    "@injectivelabs/prettier-config": "latest",
+    "@injectivelabs/stylelint-config": "latest",
     "@nuxt/module-builder": "latest",
     "@nuxt/schema": "^3.0.0-rc.12",
     "@nuxtjs/eslint-config-typescript": "latest",
