--- packages/notifications/playground/icons.json
@@ -1,1 +1,0 @@
-{"icons":["arrow-up-down","arrow","bolt","calendar","caret-down","checkbox-list","chevron-down","circle-check-border","circle-check","circle-close-border","circle-close","circle-info","close-bold","close","copy","discord","dollar-coin","download","exchange","facebook","file","gear","google","link","logo","logox","menu","monitor","nav-arrow","pen-to-square","retangle-list","sliders","telegram","triangle-down","twitter","vault","wallet/injective","wallet/keplr","wallet/ledger","wallet/metamask","wallet/torus","wallet/trezor","wallet","warn"]}
