--- packages/notifications/playground/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "components-library-playground",
-  "scripts": {
-    "generate:iconFile": "node generate.cjs",
-    "story:dev": "yarn generate:iconFile && histoire dev",
-    "story:build": "histoire build",
-    "story:preview": "histoire preview --port 4567"
-  },
-  "devDependencies": {
-    "@histoire/plugin-nuxt": "latest",
-    "@histoire/plugin-vue": "latest",
-    "@nuxtjs/tailwindcss": "latest",
-    "histoire": "latest"
-  }
-}
