+++ tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "./packages/injective-ui/tsconfig.json",
+  "compilerOptions": {
+    "types": [
+      "@vueuse/nuxt",
+      "@injectivelabs/notifications"
+    ]
+  }
+}
